move message tests to its own file
